Debugger improvement
- Passing the bus as argument in Ppu.
- Rewritten DrawInstructions
that is now using relative indices instead of a vector
filled by a thread not linked to the cpu.
- Now,
the debugger is showing the currently executed instructions instead of some random instructions
taken at a certain moment but don't know when.
- DrawInstructions is not static anymore since it using mBus.
-
The BEBUG_INFO are now set inside the cpu loop using basic types
(reducing cpu usage).
- Removes generateDebugInformations since it was not useful.
- opcodesStr is now static.